mlaas/requirements.txt
grpcio==1.20.0
grpcio-tools==1.20.0
aiohttp
redis
